Only owners can do actions with invoices

+++ invoices/core/delete.js
@@ -1,5 +1,11 @@
 'use strict';
 
-module.exports.deleteInvoice = async (id, db) => {
-    return await db.deleteInvoice(id);
+module.exports.deleteInvoice = async (id, owner, db) => {
+    const invoice = await db.getInvoiceById(id);
+    
+    if (invoice && invoice.owner === owner) {
+        return await db.deleteInvoice(id);
+    } else {
+        throw new Error('Invoice not found or you are not the owner');
+    }
 };
